WIP: spec location adjust
