+++ pkg/app/e2e/context_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package app
+package e2e
 
 import (
 	"os/exec"
@@ -20,6 +20,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
+	"github.com/tensorchord/envd/pkg/app"
 	"github.com/tensorchord/envd/pkg/home"
 	"github.com/tensorchord/envd/pkg/types"
 )
@@ -69,8 +70,8 @@ var _ = Describe("home context", func() {
 
 			It("should be able to build image with TCP context", func() {
 				args := []string{"envd.test", "--debug", "build", "--path", buildContext}
-				app := New()
-				err := app.Run(args)
+				envdApp := app.New()
+				err := envdApp.Run(args)
 				Expect(err).NotTo(HaveOccurred())
 			})
