+++ pkg/app/e2e/testdata/quick-start/demo.py
